backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.7.0
pypdf==3.17.0
python-docx==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiofiles==23.2.1
httpx==0.27.2
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]
python-decouple==3.8
aiosmtplib==3.0.1
email-validator==2.1.0
redis==5.0.1
cloudinary==1.36.0
requests==2.31.0
python-multipart==0.0.6
openai>=1.0.0